Add build.sh, remove docker-compose file
Since I cannot run a plain `docker-compose up` because of the required
`mkdir` and `xhost` commands, there's no sense in requiring
docker-compose at all. For that specific case, the `run.sh` script is
a better fit.

Add a new `build.sh` script in place of old `docker-compose build`.

Now with a build and a run script, docker-compose is not required
and we only need docker itself installed.